added and included CfgVideoOptions.hpp, removed "required" addons.
